refactor(planner): extract shared modifier chip utility
The modifier chip extraction logic was duplicated across
PlannerTreeNode, PlannerListRow, and SummoningPlannerView. This
consolidates all three copies into a single extractModifierChips
utility, adds Speed Tier and Tool Speed chip types, and updates
all consumers to import from the shared module.
